+++ internal/repository/question/pg/converter/question_option.go
@@ -7,7 +7,7 @@ import (
 
 func QuestionOptionFromPGSQL(questionOptionRepo *modelRepo.QuestionOption) *model.QuestionOption {
 	questionOptionServ := &model.QuestionOption{
-		ID:        int64(questionOptionRepo.ID),
+		ID:        questionOptionRepo.ID,
 		Text:      questionOptionRepo.Text,
 		IsCorrect: questionOptionRepo.IsCorrect,
 	}
@@ -17,7 +17,7 @@ func QuestionOptionFromPGSQL(questionOptionRepo *modelRepo.QuestionOption) *mode
 
 func NewQuestionOptionToPGSQL(questionID int64, questionOptionServ *model.NewQuestionOption) *modelRepo.NewQuestionOption {
 	newQuestionOptionRepo := &modelRepo.NewQuestionOption{
-		QuestionID: int32(questionID),
+		QuestionID: questionID,
 		Text:       questionOptionServ.Text,
 		IsCorrect:  questionOptionServ.IsCorrect,
 	}
