+++ internal/repository/question_subcategory/pg/model/question_subcategory.go
@@ -1,6 +1,6 @@
 package model
 
 type QuestionSubcategory struct {
-	QuestionID    int32
-	SubcategoryID int32
+	QuestionID    int64
+	SubcategoryID int64
 }
