+++ internal/repository/subcategory/pg/list_by_ids.go
@@ -15,7 +15,7 @@ func (r *repo) ListByIDs(ctx context.Context, ids []int64) ([]*model.Subcategory
 	idsRepo := make([]interface{}, len(ids))
 	placeholders := make([]string, len(ids))
 	for i, id := range ids {
-		idsRepo[i] = int32(id)
+		idsRepo[i] = id
 		placeholders[i] = fmt.Sprintf("$%d", i+1)
 	}
