+++ internal/service/category/tests/get_test.go
@@ -16,8 +16,8 @@ import (
 func TestGet(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
-	id := int64(gofakeit.Int64())
-	domainID := int64(gofakeit.Int64())
+	id := gofakeit.Int64()
+	domainID := gofakeit.Int64()
 	res := &model.Category{
 		ID:        id,
 		Name:      gofakeit.Name(),
